backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.2
MarkupSafe==3.0.2
packaging==24.2
pillow==11.2.1
PyMuPDF==1.25.3
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.1.0
typing_extensions==4.13.2
Werkzeug==3.1.3
